POC-2/requirements.txt
fastapi
uvicorn
httpx
pydantic
sqlalchemy
aiosqlite
google-generativeai
sendgrid
python-dotenv
